requirements.txt
pandas
pendulum
minio
apache-airflow-providers-google
apache-airflow-providers-slack
gspread
pangres